chore(FileIndexPage): move ItemLayoutButton as individual widget file
