+++ lib/file/presentation/view/file_index_page.dart
@@ -1,5 +1,3 @@
-// ignore_for_file: lines_longer_than_80_chars
-
 import 'dart:convert';
 import 'dart:math';
 import 'dart:typed_data';
@@ -346,40 +344,6 @@ class ItemGridView extends StatelessWidget {
         // TODO(yijing): show thumbnail when item is not photo
         return const SizedBox.shrink();
       },
-    );
-  }
-}
-
-class ItemLayoutButton extends StatelessWidget {
-  const ItemLayoutButton({
-    Key? key,
-    required this.onPressed,
-    required this.isSelected,
-    required this.uIconData,
-  }) : super(key: key);
-
-  final VoidCallback onPressed;
-  final bool isSelected;
-  final UIconData uIconData;
-  @override
-  Widget build(BuildContext context) {
-    return Ink(
-      width: 36,
-      height: 36,
-      decoration: BoxDecoration(
-        borderRadius: BorderRadius.circular(4),
-        color: isSelected ? UColors.ctaDark : UColors.backgroundDark,
-      ),
-      child: Center(
-        child: IconButton(
-          padding: EdgeInsets.zero,
-          onPressed: onPressed,
-          icon: UIcon(
-            uIconData,
-            color: isSelected ? UColors.white : UColors.textMed,
-          ),
-        ),
-      ),
     );
   }
 }
